python/requirements.txt
attrs==21.2.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
certifi==2021.10.8
Flask-Cors==3.0.7
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.0
distlib==0.3.3
docopt==0.6.2
filelock==3.4.0
Flask==2.0.2
Flask-JWT==0.3.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
neo4j==4.4.0
packaging==21.3
pipreqs==0.4.11
platformdirs==2.4.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.4.2
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.7
virtualenv==20.10.0
Werkzeug==2.0.2
farm-haystack[faiss]
elasticsearch